backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
python-chess==1.999
pydantic==1.10.7
httpx==0.23.3
python-dotenv==1.0.0
pytest==7.3.1
black==23.3.0
isort==5.12.0
mypy==1.3.0
supabase==1.0.3
python-multipart==0.0.6
numpy==1.24.3
asyncio==3.4.3
email-validator==2.0.0 